refactor(architecture): enforce package ownership boundaries
