+++ backend/conversionflow/src/main/java/com/nocountry/conversionflow/conversionflow_api/application/port/ConversionDispatchHandler.java
@@ -1,4 +1,4 @@
-package com.nocountry.conversionflow.conversionflow_api.service.dispatch;
+package com.nocountry.conversionflow.conversionflow_api.application.port;
 
 import com.nocountry.conversionflow.conversionflow_api.domain.enums.Provider;
