+++ backend/conversionflow/src/main/java/com/nocountry/conversionflow/conversionflow_api/infrastructure/dispatch/google/GoogleConversionDispatchHandler.java
@@ -1,9 +1,9 @@
-package com.nocountry.conversionflow.conversionflow_api.service.google;
+package com.nocountry.conversionflow.conversionflow_api.infrastructure.dispatch.google;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
+import com.nocountry.conversionflow.conversionflow_api.application.port.ConversionDispatchHandler;
 import com.nocountry.conversionflow.conversionflow_api.domain.enums.Provider;
 import com.nocountry.conversionflow.conversionflow_api.domain.event.LeadConvertedEvent;
-import com.nocountry.conversionflow.conversionflow_api.service.dispatch.ConversionDispatchHandler;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.stereotype.Service;
